+++ InformationProtection/functions/Set-MipLabel.ps1
@@ -19,7 +19,11 @@
 		The label to apply.
 	
 	.PARAMETER Path
-		Path to the file to label.
+		Path to the file(s) to label.
+
+	.PARAMETER LiteralPath
+		Path to the file(s) to label.
+		Does not interpret wildcards.
 	
 	.PARAMETER Justification
 		The reason for the label change.
@@ -78,6 +82,9 @@
 		[PsfFile]
 		$Path,
 
+		[PSFLiteralPath]
+		$LiteralPath,
+
 		[string]
 		$Justification,
 
@@ -112,7 +119,8 @@
 		}
 	}
 	process {
-		foreach ($filePath in $Path) {
+		foreach ($filePath in $Path + $LiteralPath) {
+			if (-not $filePath) { continue }
 			try { $file = Get-MipFile -Path $filePath -Session $sessionToUse -ErrorAction Stop }
 			catch {
 				Write-Error $_